orchestrator/requirements.txt
playwright==1.52.0
requests
python-dotenv
requests
beautifulsoup4
scikit-learn